refactor: move functional code to base-input
