+++ src/auro-input.js
@@ -13,241 +13,16 @@ import { repeat } from 'lit-html/directives/repeat.js';
 import { classMap } from 'lit-html/directives/class-map.js';
 import { ifDefined } from 'lit-html/directives/if-defined.js';
 import BaseInput from './base-input';
-import Cleave from 'cleave.js';
 
 // build the component class
 export default class AuroInput extends BaseInput {
 
-  constructor() {
-    super();
-
-    /**
-     * @private
-     */
-    this.showPassword = false;
-
-    /**
-     * @private
-     */
-    this.numericKeyboard = false;
-
-    /**
-     * @private
-     */
-    this.maxLength = undefined;
-
-    /**
-     * @private
-     */
-    this.inputIconName = undefined;
-
-    this.label = 'Input label is undefined';
-  }
-
   // function to define props used within the scope of this component
-  static get properties() {
-    return {
-      ...super.properties,
-      showPassword: {state: true}
-    };
-  }
-
-  /**
-   * LitElement lifecycle method. Called after the DOM has been updated.
-   * @param {Map<string, any>} changedProperties - Keys are the names of changed properties, values are the corresponding previous values.
-   * @returns {void}
-   */
-  updated(changedProperties) {
-    if (changedProperties.has('error')) {
-      this.validate();
-    }
-  }
-
-  /**
-   * Function to support show-password feature.
-   * @private
-   * @returns {void}
-   */
-  handleClickShowPassword() {
-    this.showPassword = !this.showPassword;
-    this.focus();
-  }
-
-  /**
-   * Supports show-password feature.
-   * @private
-   * @returns {string} Returns HTML for SVG, function to toggle between password icons.
-   */
-  togglePasswordIcon() {
-    if (this.showPassword) {
-      return this.hidePassword;
-    }
-
-    return this.viewPassword;
-  }
-
-  /**
-   * Supports show-password feature.
-   * @private
-   * @returns {string} HTML string, function for managing when to display the show-password icon.
-   */
-  showPasswordIcon() {
-    if (this.type === 'password') {
-      return html`
-        <button
-          class="iconButton passwordToggle"
-          @click="${this.handleClickShowPassword}"
-          tabindex="-1"
-        >${this.togglePasswordIcon()}</button>
-      `;
-    }
-
-    return null;
-  }
-
-  /**
-   * @private
-   * @returns {void}
-   */
-     handleKeyUp() {
-      const iconContainer = this.shadowRoot.querySelector('.iconContainer');
-
-      this.labelElement.classList.add('inputElement-label--sticky');
-
-      if (this.inputElement.value) {
-        iconContainer.classList.add("passwordIcon--show");
-      } else {
-        iconContainer.classList.remove("passwordIcon--show");
-      }
-    }
-
-  connectedCallback() {
-    super.connectedCallback();
-
-    // Process auto-formatting if defined
-    if (this.type) {
-      let config = null;
-
-      // Process only valid format strings
-      switch (this.type) {
-        case 'credit-card':
-          config = {
-            creditCard: true
-          };
-
-          this.numericKeyboard = true;
-
-          break;
-          // add additional supported formats and their config JSON here
-
-        default:
-          // Do nothing
-      }
-
-      // initialize CleaveJS if we have a defined config for the requested format
-      if (config) {
-        // eslint-disable-next-line no-unused-vars
-        const cleave = new Cleave(this, config);
-      }
-    }
-  }
-
-  /**
-   * Function to support credit-card feature type.
-   * @private
-   * @returns {void}
-   */
-  processCreditCard() {
-    const card = this.matchInputValueToCreditCard();
-
-    this.maxLength = card.formatLength;
-
-    if (!this.noValidate) {
-      this.customValidationMessage = card.customValidationMessage;
-    }
-
-    if (this.icon) {
-      this.inputIconName = card.cardIcon;
-    }
-  }
-
-  /**
-   * Function to support credit-card feature type.
-   * @private
-   * @returns {object} JSON with data for credit card formatting.
-   */
-  matchInputValueToCreditCard() {
-    const defaultCustomValidationMessage = 'Please enter a valid credit card number.';
-
-    // eslint-disable-next-line sort-vars, one-var
-    const creditCardTypes = [
-      {
-        name: 'American Express',
-        regex: /^(?<num>34|37)\d{0,9}/u,
-        formatLength: 17,
-        customValidationMessage: defaultCustomValidationMessage,
-        cardIcon: 'cc-amex'
-      },
-      {
-        name: 'Visa',
-        regex: /^(?<num>4)\d{0,9}/u,
-        formatLength: 19,
-        customValidationMessage: defaultCustomValidationMessage,
-        cardIcon: 'cc-visa'
-      },
-      {
-        name: 'Master Card',
-        regex: /^(?<num>5)\d{0,9}/u,
-        formatLength: 19,
-        customValidationMessage: defaultCustomValidationMessage,
-        cardIcon: 'cc-mastercard'
-      },
-      {
-        name: 'Discover Card',
-        regex: /^(?<num>6)\d{0,9}/u,
-        formatLength: 19,
-        customValidationMessage: defaultCustomValidationMessage,
-        cardIcon: 'cc-discover'
-      },
-      {
-        name: 'Alaska Airlines Visa',
-        regex: /^(?<num>4147\s34|4888\s93|4800\s11|4313\s51|4313\s07)\d{0,9}/u,
-        formatLength: 19,
-        customValidationMessage: defaultCustomValidationMessage,
-        cardIcon: 'cc-alaska'
-      }
-    ];
-
-    let type = {
-      name: 'Default Card',
-      formatLength: 19,
-      customValidationMessage: defaultCustomValidationMessage,
-      cardIcon: 'credit-card'
-    };
-
-    creditCardTypes.forEach((cardType) => {
-      if (cardType.regex.exec(this.value)) {
-        type = cardType;
-      }
-    });
-
-    return type;
-  }
-
-  /**
-   * Function to support credit-card feature type.
-   * @private
-   * @returns {string} Value for the input field after changing undefined to an empty string.
-   */
-  initializeValue() {
-    if (this.value === undefined) {
-      this.value = '';
-
-      return '';
-    }
-
-    return this.value;
-  }
+  // static get properties() {
+  //   return {
+  //     ...super.properties
+  //   };
+  // }
 
   // function that renders the HTML and CSS into  the scope of the component
   render() {
